Extract loadMessages and IntlProvider
